Pathtracer working again

+++ RaySamplesGui/TracingGui/Sources/DoubleIntSpinBox.cpp
@@ -1,6 +1,6 @@
 #include "DoubleIntSpinBox.h"
 
-void DoubleIntSpinBox::setValue(float val)
+void DoubleIntSpinBox::setValue(double val)
 {
 	QDoubleSpinBox::setValue(val);
 }
